Django_app/requirements.txt
aioredis==1.0.0
asgiref==2.3.0
asn1crypto==0.24.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
cffi==1.11.4
channels==2.0.2
channels-redis==2.0.2
constantly==15.1.0
cryptography==2.1.4
daphne==2.0.3
Django==1.11.5
django-cors-headers==2.1.0
django-leaflet==0.22.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
hiredis==0.2.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack==0.5.4
msgpack-python==0.5.4
psycopg2==2.7.3.1
pycparser==2.18
pytz==2017.2
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3